+++ docs/examples/CustomFields/DisplayIfField.js
@@ -1,5 +1,5 @@
 import React, { Children } from 'react';
-import { BaseField, nothing } from 'uniforms';
+import { BaseField } from 'uniforms';
 
 import {
   AutoForm,
@@ -11,7 +11,7 @@ import schema from './DisplayIfFieldSchema';
 // We have to ensure that there's only one child, because returning an array
 // from a component is prohibited.
 const DisplayIf = ({ children, condition }, { uniforms }) =>
-  condition(uniforms) ? Children.only(children) : nothing;
+  condition(uniforms) ? Children.only(children) : null;
 DisplayIf.contextTypes = BaseField.contextTypes;
 
 export default function ExamplesDisplayIfField() {