+++ packages/uniforms-bootstrap3/src/ErrorField.tsx
@@ -1,11 +1,9 @@
 import React from 'react';
 import classnames from 'classnames';
-import { connectField, filterDOMProps, nothing } from 'uniforms';
+import { connectField, filterDOMProps } from 'uniforms';
 
 const Error = ({ children, className, error, errorMessage, ...props }: any) =>
-  !error ? (
-    nothing
-  ) : (
+  !error ? null : (
     <div
       className={classnames('panel panel-danger text-danger', className)}
       {...filterDOMProps(props)}