+++ packages/uniforms-material/src/ErrorsField.tsx
@@ -1,14 +1,12 @@
 import FormControl from '@material-ui/core/FormControl';
 import FormHelperText from '@material-ui/core/FormHelperText';
 import React, { useContext } from 'react';
-import { context, filterDOMProps, nothing } from 'uniforms';
+import { context, filterDOMProps } from 'uniforms';
 
 const ErrorsField = ({ children, fullWidth, margin, variant, ...props }) => {
   const { error, schema } = useContext(context).uniforms;
 
-  return !error && !children ? (
-    nothing
-  ) : (
+  return !error && !children ? null : (
     <FormControl
       error={!!error}
       fullWidth={!!fullWidth}